- [upd] 下单逻辑更改、商店查询逻辑更改、远程登陆更改访问路径

+++ api-mall/src/main/java/com/hailu/cloud/api/mall/module/multiindustry/dao/StoreInformationMapper.java
@@ -15,7 +15,7 @@ public interface StoreInformationMapper {
      * @param storeSonType
      * @return
      */
-    List<StoreInformation> findStoreInformationList(@Param("storeTotalType") Integer storeTotalType, @Param("storeSonType") Integer storeSonType);
+    List<StoreInformation> findStoreInformationList(@Param("storeTotalType") Long storeTotalType, @Param("storeSonType") Long storeSonType);
 
     /**
      * 店铺详细信息
