+++ api-mall/src/main/java/com/hailu/cloud/api/mall/module/multiindustry/entity/MultiIndustryOrder.java
@@ -21,7 +21,7 @@ public class MultiIndustryOrder implements Serializable {
     /**
      * 总类型
      */
-    private Integer totalType;
+    private Long totalType;
 
     /**
      * 用户ID
